% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{clean_tpl}
\alias{clean_tpl}
\title{Clean and organize the output from TPL fuzzy matching}
\usage{
clean_tpl(x, verbose = TRUE)
}
\arguments{
\item{x}{The output from TPL fuzzy matching.}

\item{verbose}{Logical. Print detailed information?}
}
\description{
Clean and organize the output from TPL fuzzy matching
}
