% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{fix_morphotype}
\alias{fix_morphotype}
\title{Correct morphotypes}
\usage{
fix_morphotype(x, pattern = c("sp.", "spp.", "spp", "NA"),
  label = "sp.", detailed = FALSE)
}
\arguments{
\item{x}{A character vector specifying the species list, each element including
genus and specific epithet}

\item{pattern}{A character vector specifying the pattern to identify potential
morphotypes.}

\item{label}{A character string to replace morphotypes epithet.}

\item{detailed}{Logical. If TRUE returns a more detailed output.}
}
\description{
Find and fix morphotypes in a species list
}
\examples{
library(stdnames)
sp <- c("Mimosa sp.", "Eucalyptus_NA", "Yucca glauca")
fix_morphotype(x = sp)
fix_morphotype(x = sp, label = "spp.")

#change patterns to spot morphotypes
sp <- c("Mimosa unkown", "Eucalyptus_NA", "Yucca glauca")
fix_morphotype(x = sp, pattern = c("unkown", "NA"))

# No morphotypes found, x is returned back!
sp <- c("Mimosa tenuiflora", "Eucalyptus angulosa", "Yucca glauca")
fix_morphotype(x = sp, pattern = c("unkown", "NA"))

sp <- c("Mimosa tenuiflora", "Mimosa sp.", "Eucalyptus_NA", "Yucca NA",
"Acacia spp.", "Acacia ")
out <- fix_morphotype(x = sp, detailed = TRUE)
out$x
 out$morphotypes
barplot(out$types_count)
}
