% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_search.R
\name{fuzzy_genus}
\alias{fuzzy_genus}
\title{Fuzzy matching TPL genus names}
\usage{
fuzzy_genus(x, max.distance = 1, genus_list = NULL, ...)
}
\arguments{
\item{x}{A character vector with genus names to perform match.}

\item{max.distance}{Maximum distance allowed for a match.}

\item{genus_list}{The list of unique genus names. If NULL, the updated list
of names will be downloaded from The Plant List website with \link{tpl_genus}}

\item{...}{Arguments to be passed to \link{agrep}.}
}
\value{
A list with potential matches for the submitted names.
}
\description{
Try to match submitted genus name with names from The Plant List using
fuzzy maching approach.
}
