% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{parse_names}
\alias{parse_names}
\title{parse_names()}
\usage{
parse_names(x)
}
\arguments{
\item{x}{A character vector specifying the species list, each element including
genus and specific epithet}
}
\value{
A data.frame with the following components:
\itemize{
\item{\strong{original_binomial}} The original names provided in \code{x}.
\item{\strong{original_genus}} The parsed Genus name extracted from \code{x}.
\item{\strong{original_epi}} The parsed Epithet name extracted from \code{x}.
}
}
\description{
parse_names()
}
\examples{
library(stdnames)
sp <- c("Mimosa tenuiflora", "Eucalyptus_lehmannii", "Yucca glauca")
parse_names(x = sp)
}
