% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{species_data}
\format{A data frame with 19 rows and 3 variables:
\describe{
\item{\code{species}}{Plant species names}
\item{\code{var_num}}{A random numeric variable}
\item{\code{var_cat}}{A random categorical variable}
}}
\usage{
species_data
}
\description{
A example comparative dataset.
}
\keyword{datasets}
