% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{species_list}
\alias{species_list}
\title{species_list}
\format{A data frame with 19 rows and 1 variables
\describe{
\item{\code{species}}{Plant species names}
}}
\usage{
species_list
}
\description{
A example data.frame with species names.
}
\keyword{datasets}
