% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{std_names}
\alias{std_names}
\title{Standardize plant names}
\usage{
std_names(x, species_column, id_label = "gazp")
}
\arguments{
\item{x}{A data.frame with your data and one column with plant species names
that you iwsh to standardize.}

\item{species_column}{A character with the name of the column in the data.frame
that contains the plant species names that you iwsh to standardize.}

\item{id_label}{A character with the label prefix (e.g. id_) to generate a
unique identifier for the dataset.}
}
\value{
A list with the following components:
\itemize{
\item{\strong{std}} The original data.frame with standardized plant names
following The Plant List.
\item{\strong{std_names}} A data.frame with submitted and standardized plant
names following The Plant List.
\item{\strong{tpl_all}} The original data.frame with detailed information on
name standardization following The Plant List.
\item{\strong{tpl_clean}} The \code{tpl_all} data.frame after excluding
duplicated plant names.
\item{\strong{taxonomy_clean}} A taxonomic lookup table including the higher
taxonomy of each plant name in {tpl_clean}.
\item{\strong{which_fail}} The names that fail to match with The
Plant list database.
\item{\strong{which_corrected}} The names that were corrected automatically
to match The Plant List database.
\item{\strong{which_duplicated}} The names that were duplicated after
standardization (e.g removing synonyms)
\item{\strong{raw_tpl_out}} the raw output from \code{\link{TPL}}.
\item{\strong{source_data }} The source data.frame provided with unique
identifiers \code{id_names}.
\item{\strong{tpl_warnings}} Any warning during TPL fuzzy matching.
}
}
\description{
A reproducible protocol to standardize plant names from a comparative
dataset.
}
\examples{
\dontrun{
library(stdnames)

my_data <- species_data
# Stanardize species names within a comparative dataset
# First check and fix morphotypes
# check morphotypes
mf <- fix_morphotype(x = my_data$species)
mf[mf$morphotype == "yes", ]

# Replace morphotype names
my_data$species <- mf$corrected_binomial

# Run protocol to standardize plant names
out <- std_names(x = my_data, species_column = "species", id_label = "ID")

# Your dataset plus corrected names
out$std

# Only submitted/corrected names
out$std_names

# Complete output (with details on name standardization)
out$tpl_all

# After removing duplicated names
out$tpl_clean

# A higher taxonomy for the clean species list
out$taxonomy_clean

# Names that failed to match the plant list
out$which_fail

# Names that were automatically corrected
out$which_corrected

# Duplicated names after standardization
out$which_duplicated

# The complete output from TPL fuzzy matching
out$raw_tpl_out

# Source data with unique identifiers for each row in the data.frame
out$source_data
}

}
\seealso{
\code{\link{TPL}}
}
