% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_search.R
\name{tpl_genus}
\alias{tpl_genus}
\title{Updated genus names}
\usage{
tpl_genus()
}
\value{
A character vector with updated unique genus names from The Plant
List database.
}
\description{
Get the full list of genus names from The Plant List website database.
}
