% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_search.R
\name{tpl_genus_check}
\alias{tpl_genus_check}
\title{Check if genus occur in TPL}
\usage{
tpl_genus_check(genus)
}
\arguments{
\item{genus}{A character vector with genus names to check.}
}
\value{
A data.frame with the provided genus names plus if each genus occurs
in TPL.
}
\description{
Check if genus occur in TPL
}
\examples{
library(stdnames)
genus <- c("Mimosa", "Aizoom")
tpl_genus_check(genus)
}
