% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{white_space}
\alias{white_space}
\title{Remove unecessary white spaces from plant names.}
\usage{
white_space(x)
}
\arguments{
\item{x}{A character vector with plant names.}
}
\value{
The names provided after checking and removing for
Leading/Trailing and unecessary white spaces on names.
}
\description{
Remove unecessary white spaces from plant names.
}
\examples{
library(stdnames)
sp <- c("Mimosa tenuiflora ", "Yucca   glauca", " Eucalyptus angulosa")
white_space(sp)
}
