% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{std_names}
\alias{std_names}
\title{Standardize plant names}
\usage{
std_names(x, species_column, trait_columns = NULL, infra = TRUE,
  max.distance = 1, diffchar = 2, version = "1.1", verbose = FALSE)
}
\arguments{
\item{x}{A data.frame with your data and one column with plant species names
that you iwsh to standardize.}

\item{species_column}{A character with the name of the column in the data.frame
that contains the plant species names that you iwsh to standardize.}

\item{trait_columns}{A character vector with the column names in the data.frame
that contains variables (traits) you want to keep in the output.}

\item{infra}{Logical. If TRUE (default), infraspecific epithets are used to
match taxon names in TPL.}

\item{max.distance}{A number indicating the maximum distance allowed for a
match in agrep when performing corrections of spelling errors in specific
epithets. see \code{\link{agrep}}.}

\item{diffchar}{A number indicating the maximum difference between the number
of characters in corrected and original taxon names.}

\item{version}{A character vector indicating whether to connect to the newest
version of TPL (1.1) or to the older one (1.0). Defaults to "1.1".}

\item{verbose}{Logical. If TRUE detailed messages and warnings are printed.}
}
\value{
A list with the following components:
\itemize{
\item{\strong{std}} The original data.frame with standardized plant names
following The Plant List.
\item{\strong{std_names}} A data.frame with submitted and standardized plant
names following The Plant List.
\item{\strong{tpl_all}} The original data.frame with detailed information on
name standardization following The Plant List.
}
}
\description{
A reproducible protocol to standardize plant names from a comparative
dataset.
}
\details{
This function uses the functions: \code{\link{TPL}} and
\code{\link{TPLck}} from the R package \code{Taxonstand} to perform
name standardization following The Plant List database.
}
\examples{
\dontrun{
library(stdnames)

my_data <- species_data
# Stanardize species names within a comparative dataset
# First check and fix morphotypes
# check morphotypes
mf <- fix_morphotype(x = my_data$species)
mf[mf$morphotype == "yes", ]

# Replace morphotype names
my_data$species <- mf$corrected_binomial

# Run protocol to standardize plant names
out <- std_names(x = my_data, species_column = "species", id_label = "ID")

# Your dataset plus corrected names
out$std

# Only submitted/corrected names
out$std_names

# Complete output (with details on name standardization)
out$tpl_all

# After removing duplicated names
out$tpl_clean

# A higher taxonomy for the clean species list
out$taxonomy_clean

# Names that failed to match the plant list
out$which_fail

# Names that were automatically corrected
out$which_corrected

# Duplicated names after standardization
out$which_duplicated

# The complete output from TPL fuzzy matching
out$raw_tpl_out

# Source data with unique identifiers for each row in the data.frame
out$source_data
}

}
\seealso{
\code{\link{TPL}} and \code{\link{TPLck}}.
}
